07-21/requirements.txt

pycryptodome==3.15.0
